# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Contactless cashless event payment system API (Colombia market)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payouts:
    get:
      tags: [Payouts]
      operationId: listPayouts
      summary: List payouts (admin sees all; merchant_admin sees own)
      parameters:
        - name: merchantId
          in: query
          required: false
          schema:
            type: string
        - name: eventId
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of payouts.
          content:
            application/json:
              schema:
                type: object
                required: [payouts]
                properties:
                  payouts:
                    type: array
                    items:
                      $ref: '#/components/schemas/MerchantPayout'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Payouts]
      operationId: createPayout
      summary: Create a payout record for a merchant (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePayoutBody'
      responses:
        '201':
          description: Payout created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantPayout'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /payouts/{payoutId}/transactions:
    get:
      tags: [Payouts]
      operationId: getPayoutTransactions
      summary: List transactions included in a specific payout (admin or merchant_admin)
      parameters:
        - name: payoutId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Payout details with associated transactions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayoutTransactionsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Reports & Alerts ─────────────────────────────────────────────────────────
...
components:
...
  responses:
    BadRequest:
      description: Invalid request body or parameters.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Not authenticated.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: Insufficient permissions.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    TransactionLineItem:
      type: object
      required: [id, transactionLogId, productNameSnapshot, unitPriceSnapshot, unitCostSnapshot, quantity]
      properties:
        id:
          type: string
        transactionLogId:
          type: string
        productId:
          type: [string, 'null']
        productNameSnapshot:
          type: string
        unitPriceSnapshot:
          type: integer
        unitCostSnapshot:
          type: integer
        quantity:
          type: integer
    TransactionLog:
      type: object
      required: [id, idempotencyKey, braceletUid, locationId, merchantId, eventId, grossAmount, commissionAmount, netAmount, newBalance, counter, createdAt]
      properties:
        id:
          type: string
        idempotencyKey:
          type: string
        braceletUid:
          type: string
        locationId:
          type: string
        merchantId:
          type: string
        eventId:
          type: string
        grossAmount:
          type: integer
        commissionAmount:
          type: integer
        netAmount:
          type: integer
        newBalance:
          type: integer
        counter:
          type: integer
        performedByUserId:
          type: [string, 'null']
        offlineCreatedAt:
          type: [string, 'null']
          format: date-time
        syncedAt:
          type: [string, 'null']
          format: date-time
        createdAt:
          type: string
          format: date-time
        lineItems:
          type: array
          items:
            $ref: '#/components/schemas/TransactionLineItem'
...
    PayoutPaymentMethod:
      type: string
      enum: [transfer, nequi, cash, other]
    MerchantPayout:
      type: object
      required: [id, merchantId, eventId, periodFrom, periodTo, grossSales, commission, netPayout, paymentMethod, performedByUserId, paidAt, createdAt]
      properties:
        id:
          type: string
        merchantId:
          type: string
        eventId:
          type: string
        periodFrom:
          type: string
          format: date-time
        periodTo:
          type: string
          format: date-time
        grossSales:
          type: integer
        commission:
          type: integer
        netPayout:
          type: integer
        paymentMethod:
          $ref: '#/components/schemas/PayoutPaymentMethod'
        referenceNote:
          type: [string, 'null']
        performedByUserId:
          type: string
        paidAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    CreatePayoutBody:
      type: object
      required: [merchantId, eventId, periodFrom, periodTo, paymentMethod, paidAt]
      properties:
        merchantId:
          type: string
        eventId:
          type: string
        periodFrom:
          type: string
          format: date-time
        periodTo:
          type: string
          format: date-time
        paymentMethod:
          $ref: '#/components/schemas/PayoutPaymentMethod'
        referenceNote:
          type: string
        paidAt:
          type: string
          format: date-time
    PayoutTransactionsResponse:
      type: object
      required: [payout, transactions]
      properties:
        payout:
          $ref: '#/components/schemas/MerchantPayout'
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/TransactionLog'
...